third_party/blink/web_tests/wpt_internal/task-tracking/soft-navigation-heuristics/soft-navigation-detection.html

<!DOCTYPE HTML>
<html>
<head>
<meta charset="utf-8">
<title>Detect simple soft navigation.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
</head>
<body>
  <main id=main>
  <a id=link>Click me!</a>
  </main>
  <script>
    promise_test(t => {
      return new Promise(resolve => {
        const link = document.getElementById("link");
        link.addEventListener("click", e => {
          setTimeout(async () => {
            // Fetch some content
            const response = await fetch("../resources/content.json");
            const json = await response.json();
            // Change the URL
            history.pushState({}, '', "/foobar.html");
            // Add the content
            const main = document.getElementById("main");
            const content = document.createTextNode(json["content"]);
            main.appendChild(content);
            setTimeout(()=>{
              assert_equals(document.softNavigations, 1);
              resolve();
            }, 10);
          }, 0);
          e.preventDefault();
          return false;
        });
        if (test_driver) {
          test_driver.click(link);
        }
      });

    }, "Test that a soft navigation is detected");
  </script>
</body>
</html>
